relocate js/css assets with view componentfir searchable collection
