+++ app/components/shared/searchable_collection_component.rb
@@ -1,5 +1,5 @@
 class Shared::SearchableCollectionComponent < ViewComponent::Base
-  attr_accessor :form, :attribute_name, :collection, :value_method, :text_method, :hint_method, :threshold
+  attr_accessor :form, :attribute_name, :collection, :value_method, :text_method, :hint_method, :threshold, :small
 
   def initialize(form:, attribute_name:, collection:, value_method:, text_method:, hint_method:, threshold: 10)
     @form = form
