+++ app/components/shared/searchable_collection_component/searchable_collection_component.html+radiobutton.slim
@@ -9,7 +9,7 @@
     value_method,
     text_method,
     hint_method,
-    small: @small,
+    small: small,
     classes: "radiobutton-label__bold",
     legend: nil,
     hint: nil
