+++ app/components/shared/searchable_collection_component/searchable_collection_component.scss
@@ -3,11 +3,6 @@
     margin-bottom: 0;
   }
 
-  .govuk-checkboxes__item,
-  .govuk-radios__item {
-    margin-left: 10px;
-  }
-
   &.scrollable {
     .govuk-form-group {
       max-height: 194px;
@@ -19,6 +14,11 @@
     background-color: #fff;
     border: 5px solid #f3f2f1;
     padding: 0;
+
+    .govuk-checkboxes__item,
+    .govuk-radios__item {
+      margin-left: 10px;
+    }
   }
 
   .collection-component__search {
